+++ packages/common/src/index.js
@@ -1,5 +1,6 @@
 export * from './actions';
 export * from './date';
+export * from './debounce';
 export * from './dom';
 export * from './events';
 export * from './jsevents';